KAL: removed the {add,remove}{Source,Buffer} functions from KALEngine and the functions that called it
OpenAL automatically destroys buffers and sources when the device is closed, so no need to do that ourselves.
These public functions were breaking encapsulation and led to undefined behavior when called by the user.
The only change to user code is that KALEngine::instance(); must be called before using any KAL class.
The alternative would be to call instance() everytime a KAL class is instantiated but this seems rather hackish, not to mention that side effect programming is Bad.

The following functions had to be removed:
- soundNumber()
- {disable,enable}AllSound()
- get{Sound,Buffer}List()