Revert the last archive related changes from Shaan7
Reason: We would not like to have an own archive format like that because of
more reasons: it will not be supported by any external tool on any platform, it
was so simplified so that it did not even contain the most fundamental
information and it is a better practice in this case to use the standard formats
that is available. Furthermore, we do not need separate classes in this case.

KAr and K7Z kdecore libraries are under development by me to extend, implement
them and then we can gain the most benefit from those by putting somewhat to the
core module after checking out the other upstream ways like there is a similar
Qt only implementation out there. At any rate, it is something that will be
boosted with Qt developers in the hope of getting a better situation than
implementing the same functionality at more third party projects. There are at
least 5-6 projects on the tip of my tongue right away, where the same purpose is
implemented.