Fairly big speedup and proper forward declaration usage commit
The other nice outcome of this going through is that there were quite a few
unneccesary include eliminated. Hence not just an include to forward declaration
change in the desired cases.

Not sure about the performance benefits overall, but let us see and at least the
codebase got cleaned up regarding this and another quality level.