Fairly big speedup and proper forward declaration usage commit - part2
Eliminate the unneccesary commit lines from the previous try.
What is an addition is that I now also fixed the issues related to the #include
lines like:
* Missing line breaks because of the readability
* Proper order
* Once I had to change the function declaration for reference usage because it
made sense anyway and helped the forward declaration usage

All in all, the includes should now be fine inside the header files, but let us
see in the practice.