Replace the "#warning" _GNUC_ specific pragmas with "// TODO" lines
It is generally a good practice to use "// TODO" lines for items where there is
a need some further investigation. Some editors highlight it by default that
way. They can also be documented on a roadmap, todo webpage if it is needed.

One other big advantage is that the tools, like compiler, static code analyzer
and build services, build systems and so forth do not mix it up with the real
warnings.

Also, the warnings is not platform specific anymore in the compilation output which
could result false impression.