Store items in QList instead of an ordered QMap. Reason for this is that while adding items to game list, we need to know which row we are adding so that we can pass it to beginInsertRows. If we use a QMap, its very cumbersome to calculate which row the insertion will take place. With a QList, we do insertion at the end. Another QHash is needed to find out which game is at which index.
Add few convenience functions to the model to accommodate the above change.
Now, the model will NOT reset, instead it updates and inserts Game Items dynamically.
Make it possible for the view to set new rating using setData. Testing pending.
Add new enums to describe the game status in more detail. Upgradable and OwnedByUser
disabled as of now.
Make it possible for the OcsProvider to fetch details about a particular game.
Fix the QML Rating widget to properly update its values.
Fix coding style in OcsProvider.