Use const ref in foreach iteration for non-pod types
CCMAIL: jhahoneyk@gmail.com

When not using POD types (int, double, pointer, ...) you should use const & for
your foreach variables. There are two reasons for this:
1) Prevents you from the mistake of writing foreach loops that modify the list,
that is 'foreach(Foo f, list) f.a = f.b = f.c = 0;' compiles but does not
modify the contents of list
2) Saves a copy constructor call for each of the list elements

Also, according to that change I had to modify the addChildren prototype. The
second argument got a const modifier which is not a problem in this case at all
and it has been an easy fix. On the other hand, there is no need for pointer
usage since it passes just like a pointer internally, but has checks done by the
compiler.