Make use of the savable context to dirtify objects on any action
This changes the undo system lightly - the undo/redo are no longer
pure virtual, so it must be called in the reimplemented classes,
hence the large amount of changed files.

The context function now also returns a Savable pointer instead
of an Asset pointer, as that is what is actually being used here.