Do not add icons subdirs since there is no need for that
First, it is not nice to use ../* includes from a library into an application
folder which builds after the library in the stage. Also, it should work
properly even if the creator is not required for building.

Second, we do not need to add the icons dir from the touch folder since they are
compiled in resources, thus it is no use to install them if they are already
available like that.