Fix up the plugin loading issue on Windows/Mac by a one-liner change
The problem is that behind the current code, "Dir::NoDotAndDotDot" returns
nothing on Windows andaccording to the known bug report, also on Mac.

There is a closed invalid bug on Nokia regarding:
http://bugreports.qt.nokia.com/browse/QTBUG-6077

It was marked as invalid and the real solution for the issue is using
"QDir::AllEntries| QDir::NoDotAndDotDot".