Comment out the capture* audio files in the cmake file
CCMAIL: smarter@ubuntu.com
It worth it since they are not used and for the time being, the first part of
the libsndfile elimination process. Also, it can be pushed to the master branch
right now, the other patch is pending until the next alure release. It is not
that long feature, alure is now about in feature freeze state.