Fix the documentation issue again...
CCMAIL: admin@leinir.dk
License addition to the documentation folder was not enough since static code
analyzers handle it as a header file meaning that they also give that feedback
we are missing some include guard, and hence they are not proper.

There might be similar unneccesary jobs we need to do with the documentation
files in the future.