Add PinchArea to the qml player, but QtQuick 1.1 (Qt 4.7.4) is not out yet
This idea will allow proper zooming as soon as we get the Qt 4.7.4 version
released. Right now, it is implemented after a long research since it is not
properly documented and there is no proper example, tutorial for it.

Why it is an important feature is that because N9(50) Harmattan/Maemo6 devices
do not really have the stylus stick and if there are small input fields, for
instance, it is really hard to type without zooming. This phone is a different
technology, that is.