This is the easiest way of getting properly rid of the syscall krazy complain
open syscall is completely acceptable inside the input management, linux folder.
Krazy complains about it since it is not a cross-platform way of doing it even
if we do not wanna use cross-platform way here. KDE addressed this issue with
special macros internally, but since it is a pure Qt library, it is better to
avoid to make KDE as a dependency because of this syscall only.

It is not acceptable to make krazy silent in general for no real reason, but in
this special case it is the proper of going forward.