Check the touch-declarative found order prior to the KDE things
It seems this is the solution for the Harmattan, Maemo6 issue for now.
Also, this is actually the correct order meaning that we should first check the
qt simple player, qt touch player, kde simply player, et cetera. Hence this is
the proper solution for this case, but the discussion is ongoing on the cmake
mailing list since it is still a bug related to the declarative-KDE macro order.