Achievements 20: Add a threshold value to the Achievement class.
REVIEW: 103253

With this threshold value you define when the player starts to make
progress. Hidden achievements will be shown as soon as the player starts
to make progress.

Further changes:
- AchievementsManager::dependency now returns the index of the dependency

This is the last commit in my achievements branch :). After that, only the
new delegate (which I have to rewrite) is missing.