Give a more talkative test name in the unit test generation.
Just a side note: It is very odd: "make test" handles them as a separate unit,
and it cannot contain method-wise granularity. It would not be itself an issue
as such, but when I made the test wrong: it did not see which test method
failed. It is even worse when you have lots of data per test.

Most likely, the separate test case must be run according to the proper binary
name in such cases.