Introduce the modulename argument in the unit test cmake macro
It is the next step to the generic unit test cmake macro that could be used from
each module. The current issue left is that cmake macros cannot be passed by
array type arguments. The good example is GLUON_CORE_LIBS for it which is an
array type, like: A;B;C;D. Unfortunately, the cmake macro will only get the A
part of it internally.

The temporary (?) solution would be to dedicate a variable for it which needs to
be fill in before calling the macro. It is not a big deal if the macro has this
1-2 lines of documentation and the macro users read that. They might need to
read it for any usage and realization, this macro exists, anyway.