Establish a GluonMacros.cmake file and start using it
This commit puts the generic macro into a cmake file after establishing that.
It implies another step on the way, namely to set the module path correctly
inside the tests CMakeLists.txt files and including it then properly.

It eliminates a lot of duplication, triplication and so forth from the
${projectroot}/${module}/tests/CMakeLists.txt file and it just nicely makes them
as simple as possible.