Audio: Use Status enum for the states instead of more bool type variables
I have been told to use alGetSourcei for OpenAL state variables, but as I said
from the beginning, it does not work out well. It basically reports the status,
AL_INITAL, AL_PLAYING, AL_PAUSED, AL_STOPPED. The problem is that this approach
does not really work for streams. Hence it is pretty much useless for us.

The other stance I have actually implemented is to refactor the bool variable
usages for states, and I started using a status enumeration. That way the code
looks less error prone and more professional.

The other very big advantage of this solution is that, we do not need to care
about the proper state variable settings all the time anymore. It is nicely
enough if we set the proper status value from the enumeration. It also fixed
couple of bugs that I previously proposed for fixing, but this way it is cleaner
to fix them up.

There were also missing getters on the API, like isStopped, thus it has also
already been added. We got another benefit with this change meaning that we can
provide an "Initial" state to the API users which may be happening after the
construction and before the first play operation. That way the users could also
recognize this state, if they want. It is a nice addition to the API. It is also
neat since it is somewhat an easier approach for now than the initial state in
case of the OpenAL API which also considers this one after rewinding. I have not
dealt with that for now.