Add prefabinstancechildren correctly
creating with a parent is not enough here, this is a gameobject, so
we have to explicitly call addChild to ensure it's parentage is
correctly assigned. Perhaps we can automate this a little bit, but
do we have enough information during construction?

Also includes some minor cleanup, plus more debug