Implement a component based NSIS generator opportunity
Most of the groundwork is done for a component based installer. This way, it
will be much easier to ship a minimal installation process for debugging
purpose. Right now, there are two types of installation set (Full and minimal)
and there are properly separated components with display names, descriptions and
relevant groups.