Fix the filter issue of the file open dialog operation
The wrong pattern was used for filtering meaning that Windows users, for
instance, could not open the game projects manually up properly, but it now
works with this fix. The usage is now similar what other applications use and
what the Qt documentation provides as an example.

Also, the filter was not translatable which is now fixed in step since it does
not mean another line modification, just a "tr" addition to the Qt player and
"i18n" addition to the kde player.

Note: The games are not listed inside the players and no game can be played
right now even by starting it from command line with the proper arguments. This
is the next step where I need to investigate. It is not related to this patch
though.