Fix up the "no game is displayed" break in the player applications
CCMAIL: jhahoneyk@gmail.com
The issue was that there were no players displayed inside the simple qt and
kdeplayer application. It happened because of the unexpected directory path
return value from the directoryprovider.

It is now fixed by a simple modification by using the prefix install and the
directory separator for making sure. Tested properly on Linux, Windows and
Harmattan. It works fine.

You can find more information about it at the previous CPack patches, especially
this one: 422b53

Also, there was a discussion and long debugging session prior to that on the
cmake mailing list where I have been told about it:
"So the solution begins with DOT NOT USE ABSOLUTE INSTALL PATH,
give it a try and tell us how it goes."