CPack/NSIS: Handle the Windows Registry properly
1) There was a CPACK_PACKAGE_VENDOR addition since without that, CPack generated
a quite general name in the Windows Registry, like "Humanity". It is first of
all, not really talkative, second it can collide with other things if someone
else leaves it empty as well.

2) There was a registry key addition with the feasible Gluon string and version
number according to the GLUON_VERSION_STRING, thus it is not hard coded and
nice.

3) On the way I fixed up a third line since we are using lower letters for the
"set".

The result is that, we now have a proper Gluon entry if you open up your
registry by running the "regedit" command. It is something like
Computer\HKEY_LOCAL_MACHINE\DOFTWARE\Wow6432Node\Gluon\Gluon 0.72.0\(Default).
That way we can access to the install path which was a bit tricky to manage it
first since it is straight-forward in case of cmake with in the
gluon_global.h.in file, like:

After all, the next step is to make the game list and other installation prefix
related things work on Windows as well!