Get the installation path from the Registry on Windows
CCMAIL: jhahoneyk@gmail.com

After the previous patch, it was possible to implement the proper reading of the
Registry on Windows. This way, if Gluon is installed by NSIS on that platform,
it is going to pick up the proper installation path from there.

If Gluon is installed over cmake, for instance in case of development purposes
and source compilation, it will still find the cmake installation prefix in
order to work further on well.

Nevertheless, for the speed efficiency, the check is only done in case of
Windows, and the ifdef part guarantees that the performance remains actually the
same on Linux like previously.

Testing
It has been tested properly on Windows, and after installing Gluon by NSIS, the
gamelist showed up properly in the simple Qt player.

Next step
It would be nice to handle the Windows registry according to the 32/64 version
so that to make an selection which entry to pick up since they are different on
different architectures, like x86 and x86_64.