Fix the parseGDL usage in the test code.
CCMAIL: ashwin_rajeev@hotmail.com

The plain GDL string was passed in a way the gdl parser called the "wrong"
method since it was used as if it had been a filename instead of the plain gdl
string.

Test: It is tested and it now returns the parsed gluonobject list properly.

Future reference: it might be better to separate the test data from the actual
test code in order to make it modularized and easier to edit the test data, if
there is any need for that.