Fix the crash while undo after a "New Folder" action.
The immediate solution for the fix is to export the NewObjectCommand class and
add a new command to the history manager from the projectdock.

The NewObjectCommand class has been already part of the library and installed,
but it might still need some consideration in the future.

If this way is not feasible, the alternative way is to place the new folder
creation into the objectmanager and not expose the NewObjectCommand class, but
that would be a bit more intrusive change for now since we should pass the model
index between the two classes (probably signal/slot connection and relevant
argument usage).