Implement a project dock specific newCreateScene() method
It fixes the situation when you try to add a new scene from the project dock,
but the new scene is actually added to the gameproject and not the folder that
you selected. The benefit is very vivid: if you create a new game project and
you would like to add a scene to the scene folder directly, it works. I think it
is a nice feature. :)

I left the objectmanager specific newCreateScene() method there since that way,
it is still possible to create scene independently from the project dock, for
instance from the menu later.

Testing: It has been tested and works as expected. I also tried out the
undo/redo operation and that also seemed to work.