Ignore the reference for now since the parser does not actually parse that
CCMAIL: ashwin_rajeev@hotmail.com

It would be nice to have more test cases, for instance in case comments or just
simply by hard coding the object hiearchy instead of comparing the parsed with
the parsed-serialized-parsed. I think that way it could be a more intuitive and
not making any guess about parsing and serializing, if the object hiearchy is
"hard coded".

Also I would recommend to separate the test data from the test functionality,
especially in this codebase. It would make it even more modularized and much
more readable.