Use gameObject* instead of gobj*
It is better practice to use gluon/gameObject inside the gluon codebase since
gobj(ect) reminds a few people for gobject from gtk. Other than that, gobj(ect)
is not straight-forward enough in some contexts whether it is supposed to be a
gluon or game object. This way, it is clear and does not make the name too long
either.