Fix the "Empty source, OpenAL error: 40964" audio file load issue
Call the Engine::instance() explicitely in order to get the alure initialization
and update interval methods called. It is needed whenever we need to do most of
anything with openal.

Remark: singleton construction might get dangerous in the future if this class
is registered as a gluonobject since that way these things will happen before
the main entry point. In that case, we need to reconsider this design. For
further details read the commit "53ef86d" (Core/Singleton: Implement a proxy
method for getting a mutex instance for need).