Eliminate set{Max/Min}Volume and query the OpenAL for the current volume
CCMAIL: ashwin_rajeev@hotmail.com
Test cases can be removed, if there are any.

It is better to not complicate the code unneccesarily, so the max volume value
is set to a value during the source loading. It can be re-discussed, if the
need arises. Summary, there is no exposed setter for those provided by the API.

We decided that way: the volume setter/getter works properly only if there is a
source loaded, and we do not need to add any unneccesary code here either that
way.

It should also fix the test cases and the operation in general when there is a
maximum value set (in this special case: 10.0f), but the volume getter does not
return bigger value anymore, even if the value is set for a bigger one.