Fix the default maximum gain value for the source
10.0f is invalid value, and if you try to play it is not gonna work.
We did not realize this issue previously since we always used some value for the
maximum in the range of 0.0f - 1.0f, so the default 10.0f was not used
previously.

It is better to not set that anywhere, and just let it be 0.0f - 1.0f (change
the documentation in the header accordingly). AL_MIN_GAIN and AL_MAX_GAIN can
be set between 0 and 1. That is also the default setting according to the OpenAL
1.1 specification, so all good.