Remove just the last organization and application name, not everything
CCMAIL: jhahoneyk@gmail.com

Using chop is a better approach if we are about to remove certain string values
from the end of another path string. It cannot be broken by any install path
this way coming from the build system. As of performance, it is even faster
since it does not need to look up the substrings this way.

Along the way: I also fixed the place the slash where to add it to the string.
The concern is that here, paths normally return the directory path without the
leading slash, and that is actually how this API has been designed and used by
the API users heretofore.