Fix the scriptengine breakage causing non-working game examples
CCMAIL: admin@leinir.dk,fxrh@gmx.de

The comment was changed in those two lines:
-        d->appendScript( asset, className );
-        //d->buildScript();
...
+        //d->appendScript( asset, className );
+        d->buildScript();

If it is intentional, it might require further study. That was the culsprit
of the breakage. It was a bit hard and sleep deprivated time to follow git
bisect so long, but the games are at least now working nicely. :)