Move the qt player (cli util this time) to a tools/player folder
I also removed the gluon_qtplayer.desktop file in this step since desktop file
is not installed anymore, and the relevant code was removed from the
CMakeLists.txt file. Therefore, it is not worth keeping it. It can be restored
from VCS anytime, if needed.

Also, eliminate the PLAYER_ICONS_INSTALLED settings at the end of the cmake file

It is not needed anymore since there is no "Qt Basic Player" under the same
hierarchy anymore. Also, even if that was the case, the relevant cmake file of
that does not install icons anymore, thus there is no duplication in that...