Move the plasmoid player into the desktop folder for now
As discussed with Shantanu, it is not being removed yet until the relevant
active plasmoid gets really mature for usages like comment feature and other
things. The best place for it right now along with the kde desktop player is the
desktop folder. It can be removed later on purpose, if that is needed.