Move the kde basic player to the tools/player/kde directory
The player is being kept for now until there is a good way of debugging on
Windows by using the MSVC 2010 Express compiler set and environment (not gnu and
mingw). That is what most of the KDE Windows developers do after all. It is very
hard to find a proper workaround without it on Windows. Like, you need to
register for some Windows LIVE ID, which is forcy enough .. Also, even after
registration it is hackish to get it work in Visual Studio. I could even
personally manage to get the free Visual Studio registered. As for me, it is a
huge pain to get it work in any forms on Windows without it. I hope it is going
to change.

1) Removed the gluon_qtplayer.desktop file in this step since desktop
file is not installed anymore, and the relevant code was removed from the
CMakeLists.txt file. Therefore, it is not worth keeping it. It can be restored
from VCS anytime, if needed.

2) Mainwindow files are removed because of the obsolete situation of theirs.

3) The relevant CMakeLists.txt cmake file entries were removed from the "old"
CMakeLists.txt cmkae file

4) New rules added properly to the "new" CMakeLists.txt cmake files

Test: It builds fine according to the new rules.