Add cmake options for the players and refactor the player cmake file
CCMAIL: djfreestyler@gmail.com
It is now building only the desktop (kde) player by default. Therefore, all the
other players are in "OFF" state. The user or/and developer can enable other
players for building by using the relevant cmake options. It does not build the
kde basic anymore. It does not use the WITH_KDE option anymore either since if
you build the kde or plasmoid desktop player, active and so on: you obviously
need to build it with KDE if that is available on the system.

Testing: Done properly on Linux and works fine according to the cmake options
set. No regression experienced.