Delete the futile status message for ATTICA_FAILED and PLAYER_FAILED
This was added in the past long time ago, but the ATTICA_FAILED and
PLAYER_FAILED variables are not even set. Moreover, the warning message is
empty, and the other one is almost useless.

On the other hand, we have a proper messaging about the summary what options
were enabled, disabled, what could be found and what not. Hence, this removal
can safely be done.