Eliminate all the qml plasmoid related entries from the active player folder
CCMAIL: jhahoneyk@gmail.com

It involves the refactoring of the CMakeLists.txt, desktop file move, metadata
desktop file, Desktop file entry change to the new schema and so forth.

Testing: It works "properly" on Linux by running the player as
"plasmoidviewer org.kde.gluonplayer-active".

Note: Playing is not supported at all, but that is out-of-the-scope of this
straight-forward patch.