Integrate the active player into the component based installation
CCMAIL: jhahoneyk@gmail.com

Nothing special, mostly just cmake and CPack rule modifications to get it into
the component based installation. That mechanism is more or less used on Windows
by the NSIS installer generation, but it might be beneficial later for other
types, like rpm, deb and so on.