Fix the unfortunate warning in the think() method of the btcharacter class
CCMAIL: admin@leinir.dk
Fix the "qt4/QtCore/qvector.h:339:3: warning: assuming signed overflow does not
occur when assuming that (X - c) > X is always false" warning by using the gcc
Diagnostic Pragma

It was important to address because otherwise cdash keeps spamming with this
every single build (currently every single day) which is not nice. On the other
hand, if we filter these emails out, we lose the main purpose of cdash which is
being able to catch the build errors and warnings as soon as possible even
before the users start reporting those. Say, we lose the quality to a certain
excent. Hence, fixing.

Please read the following report for further details:
https://bugreports.qt.nokia.com/browse/QTBUG-5915