graphics: Remove Manager, add Backend::currentBackend() and Defaults and use ResourceManager where necessary
Not really sure how to do this cleanly so this is just one big patch
that changes a bit of the fundaments. This removes the awkward Manager
class - yay! Resource management functionality is now in core so we use
that and split the rest of the functionality into the backend creation
bits - which are now in backend as they should be - and the defaults
bits which are now handled by Defaults. The currentWorld concept is
removed and Worlds are now just resources that are managed by the
ResourceManager. For the components, we should use a world per scene
using the scene name as identifier.